Query vessel's charter status from vessels table
