+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/repositories/LastPositionRepository.kt
@@ -17,8 +17,6 @@ interface LastPositionRepository {
         isValidated: Boolean,
     )
 
-    fun findUnderCharterForVessel(vesselIdentifier: VesselIdentifier, value: String): Boolean
-
     // For test purpose
     fun deleteAll()
 }
