+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/use_cases/reporting/GetAllCurrentReportings.kt
@@ -7,7 +7,7 @@ import fr.gouv.cnsp.monitorfish.domain.entities.reporting.Reporting
 import fr.gouv.cnsp.monitorfish.domain.entities.reporting.ReportingType
 import fr.gouv.cnsp.monitorfish.domain.entities.reporting.filters.ReportingFilter
 import fr.gouv.cnsp.monitorfish.domain.entities.vessel.VesselIdentifier
-import fr.gouv.cnsp.monitorfish.domain.repositories.LastPositionRepository
+import fr.gouv.cnsp.monitorfish.domain.repositories.VesselRepository
 import fr.gouv.cnsp.monitorfish.domain.repositories.ReportingRepository
 import fr.gouv.cnsp.monitorfish.domain.use_cases.control_units.GetAllControlUnits
 import org.slf4j.Logger
@@ -16,7 +16,7 @@ import org.slf4j.LoggerFactory
 @UseCase
 class GetAllCurrentReportings(
     private val reportingRepository: ReportingRepository,
-    private val lastPositionRepository: LastPositionRepository,
+    private val vesselRepository: VesselRepository,
     private val getAllControlUnits: GetAllControlUnits,
 ) {
     private val logger: Logger = LoggerFactory.getLogger(GetAllCurrentReportings::class.java)
@@ -40,7 +40,7 @@ class GetAllCurrentReportings(
                             require(it.internalReferenceNumber != null) {
                                 "The fields 'internalReferenceNumber' must be not null when the vessel identifier is INTERNAL_REFERENCE_NUMBER."
                             }
-                            lastPositionRepository.findUnderCharterForVessel(
+                            vesselRepository.findUnderCharterForVessel(
                                 it.vesselIdentifier,
                                 it.internalReferenceNumber,
                             )
@@ -50,14 +50,14 @@ class GetAllCurrentReportings(
                             require(it.ircs != null) {
                                 "The fields 'ircs' must be not null when the vessel identifier is IRCS."
                             }
-                            lastPositionRepository.findUnderCharterForVessel(it.vesselIdentifier, it.ircs)
+                            vesselRepository.findUnderCharterForVessel(it.vesselIdentifier, it.ircs)
                         }
 
                         VesselIdentifier.EXTERNAL_REFERENCE_NUMBER -> {
                             require(it.externalReferenceNumber != null) {
                                 "The fields 'externalReferenceNumber' must be not null when the vessel identifier is EXTERNAL_REFERENCE_NUMBER."
                             }
-                            lastPositionRepository.findUnderCharterForVessel(
+                            vesselRepository.findUnderCharterForVessel(
                                 it.vesselIdentifier,
                                 it.externalReferenceNumber,
                             )
