+++ backend/src/test/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/database/repositories/JpaLastPositionRepositoryITests.kt
@@ -179,17 +179,4 @@ class JpaLastPositionRepositoryITests : AbstractDBTests() {
         assertThat(previousLastPosition.alerts).contains("MISSING_FAR_ALERT")
         assertThat(lastPosition.reportings).hasSize(0)
     }
-
-    @Test
-    @Transactional
-    fun `findUnderCharterForVessel Should get the underCharter field of a vessel`() {
-        // When
-        val underCharter = jpaLastPositionRepository.findUnderCharterForVessel(
-            VesselIdentifier.INTERNAL_REFERENCE_NUMBER,
-            "ABC000180832",
-        )
-
-        // Then
-        assertThat(underCharter).isTrue
-    }
 }
